requirements.txt
argparse
faker
python-docx
pandas
fpdf
openpyxl